refactor: adding lambda client
