+++ runtime/lib/HTTPClient/function-lambda/interface.ts
@@ -1,0 +1,5 @@
+export interface FunctionLambdaRequest {
+  code: string;
+  variables: Record<string, unknown>;
+  enableLog?: boolean;
+}
